requirements.txt
geocoder
folium
scapy
python-nmap
streamlit
streamlit_folium
socket